+++ NEWS.md
@@ -1,14 +1,18 @@
 
+# cdata 0.1.6 2017/10/12
 
-'cdata' 0.1.5 2017/07/04
+ * work around empty keyset issues.
+ * add column control.
+
+# cdata 0.1.5 2017/07/04
 
  * Allow NA in key columns.
  * Add optional class annotation when moving values to rows.
 
-'cdata' 0.1.1 2017/05/05
+# cdata 0.1.1 2017/05/05
 
  * ungroup before calculating distinct.
 
-'cdata' 0.1.0 2017/03/28
+# cdata 0.1.0 2017/03/28
 
  * First release.